Integer coordinates to doubles

+++ android/src/com/mygdx/game/Analysis/DatabaseData.java
@@ -36,7 +36,7 @@ public class DatabaseData implements Data {
         DebugLog.log(String.valueOf(nnCoordinate.x));
         DebugLog.log(String.valueOf(nnCoordinate.y));
 
-        return new Vector3(nnCoordinate.x, nnCoordinate.y, 0);
+        return new Vector3((float) nnCoordinate.x, (float) nnCoordinate.y, 0);
     }
 
     /**
